# Wrap multi-return APIs using structs: avoid heap escape