* For https://github.com/golang-fips/openssl/issues/128
* Replaces the optimization removed in https://github.com/golang-fips/openssl/pull/129
* I prefer this approach over https://github.com/golang-fips/openssl/pull/130, because this approach more faithfully passes data to/from the OpenSSL API to deal with in Go code.

```
goos: linux
goarch: amd64
pkg: github.com/golang-fips/openssl/v2
cpu: Intel(R) Core(TM) i7-6700 CPU @ 3.40GHz
       │   old.txt   │        new-struct.txt        │
       │   sec/op    │   sec/op     vs base         │
ECDH-8   104.1µ ± 3%   103.8µ ± 2%  ~ (p=0.878 n=8)

       │  old.txt   │          new-struct.txt           │
       │    B/op    │    B/op     vs base               │
ECDH-8   40.00 ± 0%   32.00 ± 0%  -20.00% (p=0.000 n=8)

       │  old.txt   │          new-struct.txt           │
       │ allocs/op  │ allocs/op   vs base               │
ECDH-8   2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=8)
```